+++ src/index.js
@@ -1,3 +1,5 @@
+import "babel-polyfill";
+import 'raf/polyfill';
 import React from 'react';
 import addons from '@storybook/addons';